+++ apps/api-admin/app/Models/User.php
@@ -11,7 +11,7 @@ use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 
-#[Fillable(['name', 'email', 'password', 'avatar_path', 'is_platform_admin'])]
+#[Fillable(['name', 'email', 'password', 'avatar_path', 'username', 'phone', 'bio', 'location', 'website', 'is_platform_admin'])]
 #[Hidden(['password', 'remember_token'])]
 class User extends Authenticatable
 {